requirements.txt
fake-useragent==2.2.0
httpx==0.28.1
loguru==0.7.3
paramiko==4.0.0
pydantic==2.13.3
python-telegram-bot==22.7
pyyaml==6.0.3
tzdata